# Conflicts:
#	sustajn_admin/lib/auth/screens/dashboard_screen.dart
#	sustajn_admin/lib/constants/network_urls.dart
#	sustajn_admin/lib/constants/string_utils.dart

+++ sustajn_admin/lib/transactions/models/transaction_subscription_data.dart
@@ -29,7 +29,7 @@ class SubscriptionData {
 
 class SubscriptionDataList {
   String? monthYear;
-  int? monthWiseTotalSusbcriptionAmount;
+  double? monthWiseTotalSusbcriptionAmount;
   List<DateWiseSubscription>? dateWiseSubscription;
 
   SubscriptionDataList(
@@ -67,7 +67,7 @@ class DateWiseSubscription {
   String? userType;
   String? restaurantAddress;
   String? planType;
-  int? amount;
+  double? amount;
   String? formattedDate;
 
   DateWiseSubscription(
@@ -85,7 +85,7 @@ class DateWiseSubscription {
     userType = json['userType']?? '';
     restaurantAddress = json['restaurantAddress']?? '';
     planType = json['planType']?? '';
-    amount = json['amount']?? 0;
+    amount = json['amount']?? 0.0;
     formattedDate = json['formattedDate']?? '';
   }
